Modificaciones en el modulo de usuarios

+++ src/main/java/com/hortalsoft/users/domain/port/usecase/association/DeleteAssociationService.java
@@ -3,34 +3,39 @@ package com.hortalsoft.users.domain.port.usecase.association;
 
 import com.hortalsoft.crosscutting.util.ExceptionHortalsoft;
 import com.hortalsoft.crosscutting.util.Layer;
-import com.hortalsoft.products.domain.mapper.MapperDomainToEntity;
 import com.hortalsoft.users.domain.domain.Association;
-import com.hortalsoft.users.domain.entity.AssociationEntity;
 import com.hortalsoft.users.domain.port.input.association.DeleteAssociationUseCase;
 import com.hortalsoft.users.domain.repository.AssociationRepository;
+import com.hortalsoft.users.domain.repository.UserRepository;
+import com.hortalsoft.users.domain.specification.user.UserExistByIdSpec;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 
 @Service
 @Transactional
 public class DeleteAssociationService implements DeleteAssociationUseCase {
 
     private final AssociationRepository associationRepository;
-
+    private final UserRepository userRepository;
     private static final Layer layer = Layer.DOMAIN;
-    MapperDomainToEntity<Association, AssociationEntity> mapperDomainToEntity = new MapperDomainToEntity<>();
 
     @Autowired
-    public DeleteAssociationService(AssociationRepository associationRepository) {
+    public DeleteAssociationService(AssociationRepository associationRepository, UserRepository userRepository) {
         this.associationRepository = associationRepository;
+        this.userRepository = userRepository;
     }
 
     @Override
     public void execute(Association domain) {
         try {
-            AssociationEntity entity = mapperDomainToEntity.mapToEntity(domain, AssociationEntity.class);
-            associationRepository.deleteById(domain.getId());
+            UserExistByIdSpec userExistByIdSpec = new UserExistByIdSpec(userRepository);
+            if (userExistByIdSpec.isSatisfiedBy(domain.getId())){
+                associationRepository.deleteById(domain.getId());
+            } else{
+                throw new ExceptionHortalsoft("Usuario no encontrada", 6001, layer);
+            }
 
         } catch (ExceptionHortalsoft exceptionHortalsoft) {
             throw exceptionHortalsoft;
